+++ components/puzzles/layout-buttons.tsx
@@ -5,7 +5,7 @@ import PuzzlesDropdown from "@components/puzzles/dropdown";
 export interface LayoutButtonsProps {
   isGrid: boolean;
   puzzlesCount: number;
-  setView: ({ gridView }: { gridView: boolean }) => void;
+  setView: (gridLayout: "grid" | "list") => void;
 }
 
 const LayoutButtons = ({
@@ -16,24 +16,24 @@ const LayoutButtons = ({
   return (
     <div className="flex mt-8">
       <button
-        onClick={() => setView({ gridView: true })}
-        aria-label="set grid view"
-        className={clsx(
-          "border mr-2 bg-white/10 p-2 rounded-md transition-all duration-200",
-          isGrid ? "border-white/20" : "border-transparent text-gray-400"
-        )}
-      >
-        <ViewGridIcon className="h-5 w-5" aria-hidden="true" />
-      </button>
-      <button
-        onClick={() => setView({ gridView: false })}
+        onClick={() => setView("list")}
         aria-label="set list view"
         className={clsx(
-          "border bg-white/10 p-2 rounded-md transition-all duration-200	",
+          "border mr-2 bg-white/10 p-2 rounded-md transition-all duration-200",
           !isGrid ? "border-white/20" : "border-transparent text-gray-400"
         )}
       >
         <ViewListIcon className="h-5 w-5" aria-hidden="true" />
+      </button>
+      <button
+        onClick={() => setView("grid")}
+        aria-label="set grid view"
+        className={clsx(
+          "border bg-white/10 p-2 rounded-md transition-all duration-200 hover:bg-white/20",
+          isGrid ? "border-white/20" : "border-transparent text-gray-400"
+        )}
+      >
+        <ViewGridIcon className="h-5 w-5" aria-hidden="true" />
       </button>
       <PuzzlesDropdown currentCount={puzzlesCount || 8} />
     </div>
